[WPE] Unreviewed test gardening. Remove test entries referencing bug209859.
The bug was keeping tracking of several WebXR tests. The bug is
closed and all the tests referencing that bug are marked as passing.
WebXR feature is only enabled in WPE, all other ports skipping WebXR
tests. In conclusion, these entries in test expectations are no longer
necessary.

* platform/wpe/TestExpectations:

Identifier: 232408@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@270745 268f45cc-cd09-0410-ab3c-d52691b4dbfc